Split project into separate crates
Using a workspace didn't really work for me because I need
different targets for each package.

* arduino-expander is using the avr-*.json targets, which are no_std
* serial-gui will be using the native target and uses std
* gpio_actions is a library and can use any target, but is always no_std

Maybe there's some way to encapsulate this behaviour in build
configurations, but for now, this is easier.

It's a little annoying because I have to open one crate in its own
window for rust-analyzer to work properly, but whatever.